requirements.txt
pip==24.1.1
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
itsdangerous==2.2.0
packaging==24.1
pymongo==4.7.3
python-dotenv==1.0.1
waitress==3.0.0
Werkzeug==3.0.3